Better sudo / package load-out
The base image now puts the named user in the staff group, and sets up
sudoers(.d) to allow sudo without a password.  Needless to say, this is
intended for local use, when you trust yourself not to be bad.

It also adds a standard set of network-y tools to the base package
installation load-out, so that you can dig, ping, cat the nets, etc.